task/backend/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
django==5.1.3; python_version >= '3.10'
django-cors-headers==4.6.0; python_version >= '3.9'
django-ninja==1.3.0; python_version >= '3.7'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
fire==0.7.0
markdown-it-py==3.0.0; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
numpy==2.1.3; python_version == '3.11'
pandas==2.2.3; python_version >= '3.9'
pydantic==2.9.2; python_version >= '3.8'
pydantic-core==2.23.4; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyjwt==2.10.0; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
rav==0.0.9; python_version >= '3.7'
rest-framework-simplejwt==0.0.2; python_version >= '3.0'
rich==13.9.4; python_full_version >= '3.8.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.2; python_version >= '3.8'
termcolor==2.5.0; python_version >= '3.9'
tradingeconomics==4.4.0
typing-extensions==4.12.2; python_version < '3.13'
tzdata==2024.2; sys_platform == 'win32'
websocket-client==1.8.0; python_version >= '3.8'
